backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0
nltk==3.8.1
spacy==3.7.2

# Vector database and embeddings
faiss-cpu==1.7.4
sentence-transformers==2.2.2
numpy==1.24.3
scikit-learn==1.3.2

# Hugging Face
transformers==4.35.2
torch==2.1.1
accelerate==0.24.1
datasets==2.14.6

# Web search and scraping
requests==2.31.0
beautifulsoup4==4.12.2
googlesearch-python==1.2.3
duckduckgo-search==3.9.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Utilities
pydantic==2.5.0
httpx==0.25.2
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Logging and monitoring
loguru==0.7.2
prometheus-client==0.19.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0